refactor(portal): extract shared tRPC/auth provider setup
Both clinician and patient portals had identical tRPC client setup,
auth context/provider, and root Providers wrapper. Move these to a
new @carebridge/portal-shared package and re-export from each portal
so existing import paths continue to work unchanged.

Closes #36
