+++ apps/clinician-portal/app/providers.tsx
@@ -1,36 +1,3 @@
-// TODO: This Providers component is duplicated in patient-portal/app/providers.tsx.
-// Extract to a shared @carebridge/ui package when adding a third portal.
 "use client";
 
-import { useState } from "react";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { trpc, getTRPCLinks } from "@/lib/trpc";
-import { AuthProvider } from "@/lib/auth";
-
-export function Providers({ children }: { children: React.ReactNode }) {
-  const [queryClient] = useState(
-    () =>
-      new QueryClient({
-        defaultOptions: {
-          queries: {
-            staleTime: 30_000,
-            retry: 1,
-          },
-        },
-      }),
-  );
-
-  const [trpcClient] = useState(() =>
-    trpc.createClient({
-      links: getTRPCLinks(),
-    }),
-  );
-
-  return (
-    <trpc.Provider client={trpcClient} queryClient={queryClient}>
-      <QueryClientProvider client={queryClient}>
-        <AuthProvider>{children}</AuthProvider>
-      </QueryClientProvider>
-    </trpc.Provider>
-  );
-}
+export { Providers } from "@carebridge/portal-shared/providers";
