+++ apps/clinician-portal/src/lib/trpc.ts
@@ -1,56 +1,3 @@
 "use client";
 
-import { createTRPCReact, httpBatchLink } from "@trpc/react-query";
-import { createTRPCClient, httpBatchLink as vanillaHttpBatchLink } from "@trpc/client";
-import type { AppRouter } from "@carebridge/api-gateway/src/router.js";
-import { createAuthFetch } from "./session-expiry.js";
-
-/**
- * React Query-integrated tRPC hooks for use inside components.
- */
-export const trpc = createTRPCReact<AppRouter>();
-
-/**
- * Returns the session token stored in localStorage (if any).
- */
-function getSessionToken(): string | undefined {
-  if (typeof window === "undefined") return undefined;
-  return localStorage.getItem("carebridge_session") ?? undefined;
-}
-
-/**
- * Fetch wrapper that intercepts 401 responses and triggers session expiry.
- */
-const authFetch = typeof window !== "undefined" ? createAuthFetch() : fetch;
-
-/**
- * Create links array for the tRPC React provider.
- */
-export function getTRPCLinks() {
-  return [
-    httpBatchLink({
-      url: "http://localhost:4000/trpc",
-      fetch: authFetch,
-      headers() {
-        const token = getSessionToken();
-        return token ? { Authorization: `Bearer ${token}` } : {};
-      },
-    }),
-  ];
-}
-
-/**
- * Vanilla (non-React) tRPC client for use outside of components (e.g. login page actions).
- */
-export const trpcVanilla = createTRPCClient<AppRouter>({
-  links: [
-    vanillaHttpBatchLink({
-      url: "http://localhost:4000/trpc",
-      fetch: authFetch,
-      headers() {
-        const token = getSessionToken();
-        return token ? { Authorization: `Bearer ${token}` } : {};
-      },
-    }),
-  ],
-});
+export { trpc, getTRPCLinks, trpcVanilla } from "@carebridge/portal-shared/trpc";
